+++ PSVirtualEnv/Classes/ModuleValidation.ps1
@@ -29,5 +29,3 @@ Class ValidateVirtualEnvLocal: IValidateSetValuesGenerator {
         return [String[]] (ValidateVirtualEnvLocalDirectories)
     }
 }
-
-