+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.5.6](https://github.com/wbrandenburger/PSVirtualEnv/tree/0.5.6) (2019-09-21)
+
+**Implemented enhancements:**
+
+- Added function `Invoke-VirtualEnv` for invoking commands in a specific virtual environment.
+- Enabled installation of more than one package via function `Install-VirtualEnv`
+- Enabled autocompletion of folders with local binaries of python packages for offline installation via `Install-VirtualEnv`
+
+**Fixed bugs:**
+
+- Incorrect python path in function `Find-Python` was generated, when a virtual environment was activated.
+
 ## [0.5.5](https://github.com/wbrandenburger/PSVirtualEnv/tree/0.5.5) (2019-09-18)
 
 **Fixed bugs:**
@@ -21,6 +33,8 @@
 - Fixed problems with missing configuration file
 
 ## [0.5.1](https://github.com/wbrandenburger/PSVirtualEnv/tree/0.5.1) (2019-09-15)
+
+**Implemented enhancements:**
 
 - Added function `Edit-VirtualEnvConfig`.
 - Referenced fields in configuration files with a reference to another fields can be evaluated.